refactor: PR 반영 - Players 도메인 분리
